+++ source/robot_model/include/robot_model/Model.hpp
@@ -331,6 +331,24 @@ public:
   const pinocchio::Model& get_pinocchio_model() const;
 
   /**
+   * @brief Getter of the pinocchio model
+   * @return the pinocchio model
+   */
+  pinocchio::Model& get_pinocchio_model();
+
+  /**
+   * @brief Getter of the pinocchio data object tied to the model
+   * @return the pinocchio data object
+   */
+  const pinocchio::Data& get_pinocchio_data() const;
+
+  /**
+   * @brief Getter of the pinocchio data object tied to the model
+   * @return the pinocchio data object
+   */
+  pinocchio::Data& get_pinocchio_data();
+
+  /**
    * @brief Compute the Jacobian from a given joint state at the frame given in parameter
    * @param joint_positions containing the joint positions of the robot
    * @param frame name of the frame at which to compute the Jacobian, if empty computed for the last frame
@@ -628,5 +646,17 @@ inline void Model::set_gravity_vector(const Eigen::Vector3d& gravity) {
 
 inline const pinocchio::Model& Model::get_pinocchio_model() const {
   return this->robot_model_;
+}
+
+inline pinocchio::Model& Model::get_pinocchio_model() {
+  return this->robot_model_;
+}
+
+inline const pinocchio::Data& Model::get_pinocchio_data() const {
+  return this->robot_data_;
+}
+
+inline pinocchio::Data& Model::get_pinocchio_data() {
+  return this->robot_data_;
 }
 }// namespace robot_model